refactor(engine): extract AbstractRequestManagingLiftController base class
NaiveLiftController and DirectionalScanLiftController duplicated request
tracking, idle/home-parking, and out-of-service lifecycle logic. Move the
shared state and behavior into a new abstract base class so each
controller only implements its scheduling strategy in decideNextAction.

Closes #603
