+++ src/main/java/com/liftsimulator/engine/NaiveLiftController.java
@@ -1,19 +1,13 @@
 package com.liftsimulator.engine;
 
 import com.liftsimulator.domain.Action;
-import com.liftsimulator.domain.CarCall;
 import com.liftsimulator.domain.DoorState;
-import com.liftsimulator.domain.HallCall;
 import com.liftsimulator.domain.IdleParkingMode;
 import com.liftsimulator.domain.LiftRequest;
 import com.liftsimulator.domain.LiftState;
 import com.liftsimulator.domain.LiftStatus;
 import com.liftsimulator.domain.RequestState;
 
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.TreeSet;
@@ -39,20 +33,7 @@ import java.util.stream.Collectors;
  *
  * Optimization is out of scope; this focuses on correctness first.
  */
-public final class NaiveLiftController implements RequestManagingLiftController {
-    private static final int DEFAULT_HOME_FLOOR = 0;
-    private static final int DEFAULT_IDLE_TIMEOUT_TICKS = 5;
-    private static final IdleParkingMode DEFAULT_IDLE_PARKING_MODE = IdleParkingMode.PARK_TO_HOME_FLOOR;
-
-    private final Map<Long, LiftRequest> requestsById = new HashMap<>();
-    private final Set<LiftRequest> activeRequests = new HashSet<>();
-    private final Set<LiftRequest> completedRequests = new HashSet<>();
-    private final int homeFloor;
-    private final int idleTimeoutTicks;
-    private final IdleParkingMode idleParkingMode;
-    private Long idleStartTick;
-    private boolean parkingInProgress;
-    private boolean outOfService;
+public final class NaiveLiftController extends AbstractRequestManagingLiftController {
 
     public NaiveLiftController() {
         this(DEFAULT_HOME_FLOOR, DEFAULT_IDLE_TIMEOUT_TICKS, DEFAULT_IDLE_PARKING_MODE);
@@ -63,85 +44,7 @@ public final class NaiveLiftController implements RequestManagingLiftController 
     }
 
     public NaiveLiftController(int homeFloor, int idleTimeoutTicks, IdleParkingMode idleParkingMode) {
-        if (idleTimeoutTicks < 0) {
-            throw new IllegalArgumentException("idleTimeoutTicks must be >= 0");
-        }
-        if (idleParkingMode == null) {
-            throw new IllegalArgumentException("idleParkingMode must not be null");
-        }
-        this.homeFloor = homeFloor;
-        this.idleTimeoutTicks = idleTimeoutTicks;
-        this.idleParkingMode = idleParkingMode;
-    }
-
-    /**
-     * Adds a car call (destination request from inside the lift).
-     *
-     * @param carCall the car call to add
-     */
-    public void addCarCall(CarCall carCall) {
-        if (outOfService) {
-            return;
-        }
-        LiftRequest request = LiftRequest.carCall(carCall.destinationFloor());
-        request.transitionTo(RequestState.QUEUED);
-        trackRequest(request);
-    }
-
-    /**
-     * Adds a hall call (request from a floor to go up or down).
-     *
-     * @param hallCall the hall call to add
-     */
-    public void addHallCall(HallCall hallCall) {
-        if (outOfService) {
-            return;
-        }
-        LiftRequest request = LiftRequest.hallCall(hallCall.floor(), hallCall.direction());
-        request.transitionTo(RequestState.QUEUED);
-        trackRequest(request);
-    }
-
-    /**
-     * Adds a lift request directly.
-     *
-     * @param request the lift request to add
-     */
-    public void addRequest(LiftRequest request) {
-        if (outOfService) {
-            return;
-        }
-        if (request.getState() == RequestState.CREATED) {
-            request.transitionTo(RequestState.QUEUED);
-        }
-        trackRequest(request);
-    }
-
-    /**
-     * Cancels a lift request by its ID.
-     * The request will be transitioned to CANCELLED state and removed from the queue.
-     * This method is safe to call for requests in any state.
-     *
-     * @param requestId the ID of the request to cancel
-     * @return true if the request was found and cancelled, false if not found or already terminal
-     */
-    public boolean cancelRequest(long requestId) {
-        LiftRequest request = requestsById.get(requestId);
-        if (request == null) {
-            return false;
-        }
-
-        // If already terminal (completed or cancelled), nothing to do.
-        if (request.isTerminal()) {
-            return false;
-        }
-
-        // Transition to cancelled state.
-        request.transitionTo(RequestState.CANCELLED);
-
-        archiveRequest(request);
-
-        return true;
+        super(homeFloor, idleTimeoutTicks, idleParkingMode);
     }
 
     /**
@@ -152,7 +55,7 @@ public final class NaiveLiftController implements RequestManagingLiftController 
      * @param floor the floor to complete requests for
      */
     private void completeRequestsForFloor(int floor) {
-        Set<LiftRequest> requestsToComplete = activeRequests.stream()
+        Set<LiftRequest> requestsToComplete = getActiveRequests().stream()
                 .filter(request -> !request.isTerminal() && request.getTargetFloor() == floor)
                 .collect(Collectors.toSet());
 
@@ -175,18 +78,9 @@ public final class NaiveLiftController implements RequestManagingLiftController 
      * @return true if there is an active request for this floor
      */
     private boolean hasRequestForFloor(int floor) {
-        return activeRequests.stream()
+        return getActiveRequests().stream()
                 .filter(request -> !request.isTerminal())
                 .anyMatch(request -> request.getTargetFloor() == floor);
-    }
-
-    /**
-     * Gets all active (non-terminal) requests.
-     *
-     * @return set of active requests
-     */
-    private Set<LiftRequest> getActiveRequests() {
-        return Collections.unmodifiableSet(activeRequests);
     }
 
     /**
@@ -240,25 +134,6 @@ public final class NaiveLiftController implements RequestManagingLiftController 
                 .filter(request -> request.getTargetFloor() == targetFloor)
                 .filter(request -> request.getState() == RequestState.ASSIGNED)
                 .forEach(request -> request.transitionTo(RequestState.SERVING));
-    }
-
-    private void resetIdleTracking() {
-        idleStartTick = null;
-        parkingInProgress = false;
-    }
-
-    private boolean shouldTrackIdle(LiftState currentState) {
-        return currentState.getStatus() == LiftStatus.IDLE && currentState.getDoorState() == DoorState.CLOSED;
-    }
-
-    private Action moveTowardHome(int currentFloor) {
-        if (currentFloor < homeFloor) {
-            return Action.MOVE_UP;
-        }
-        if (currentFloor > homeFloor) {
-            return Action.MOVE_DOWN;
-        }
-        return Action.IDLE;
     }
 
     @Override
@@ -337,90 +212,10 @@ public final class NaiveLiftController implements RequestManagingLiftController 
         }
 
         if (!hasActiveRequests) {
-            if (parkingInProgress) {
-                if (currentFloor == homeFloor) {
-                    parkingInProgress = false;
-                    idleStartTick = currentTick;
-                    return Action.IDLE;
-                }
-                return moveTowardHome(currentFloor);
-            }
-
-            if (shouldTrackIdle(currentState)) {
-                Long trackedIdleStartTick = idleStartTick;
-                if (trackedIdleStartTick == null) {
-                    trackedIdleStartTick = currentTick;
-                    idleStartTick = currentTick;
-                }
-                long idleTicks = currentTick - trackedIdleStartTick;
-                if (idleTicks >= idleTimeoutTicks && currentFloor != homeFloor
-                        && idleParkingMode == IdleParkingMode.PARK_TO_HOME_FLOOR) {
-                    parkingInProgress = true;
-                    return moveTowardHome(currentFloor);
-                }
-            } else {
-                idleStartTick = null;
-            }
+            return handleIdleParking(currentState, currentTick);
         }
 
         // No requests, stay idle.
         return Action.IDLE;
-    }
-
-    /**
-     * Takes the lift out of service by cancelling all active requests.
-     * This method cancels all non-terminal requests (QUEUED, ASSIGNED, SERVING).
-     * After calling this method, the lift should be transitioned to OUT_OF_SERVICE
-     * state via the SimulationEngine.
-     *
-     * The lift will:
-     * - Stop at the nearest floor if moving
-     * - Open and close doors if at a floor
-     * - Cancel all pending requests
-     */
-    public void takeOutOfService() {
-        outOfService = true;
-        // Cancel all active (non-terminal) requests.
-        Set<LiftRequest> activeRequests = new HashSet<>(getActiveRequests());
-        for (LiftRequest request : activeRequests) {
-            if (!request.isTerminal()) {
-                request.transitionTo(RequestState.CANCELLED);
-                archiveRequest(request);
-            }
-        }
-
-        // Reset idle tracking and parking state.
-        resetIdleTracking();
-    }
-
-    /**
-     * Prepares the lift to return to service.
-     * After calling this method, the lift should be transitioned from OUT_OF_SERVICE
-     * to IDLE state via the SimulationEngine.
-     */
-    public void returnToService() {
-        outOfService = false;
-        // Reset idle tracking when returning to service.
-        resetIdleTracking();
-    }
-
-    /**
-     * Gets all requests (for testing purposes).
-     *
-     * @return a copy of all requests
-     */
-    public Set<LiftRequest> getRequests() {
-        return new HashSet<>(activeRequests);
-    }
-
-    private void archiveRequest(LiftRequest request) {
-        activeRequests.remove(request);
-        completedRequests.add(request);
-        requestsById.remove(request.getId());
-    }
-
-    private void trackRequest(LiftRequest request) {
-        activeRequests.add(request);
-        requestsById.put(request.getId(), request);
     }
 }
